+++ include/GL/IGLDrawWork.h
@@ -12,7 +12,7 @@ public:
     virtual void onDraw(GLTexture* src, GLvboBuffer* vs, GLvboBuffer* ts) = 0;
     
     /*All parameters is in [0,1), set self parameters by these, if parameters is NULL, don't do anything just return the number of parameters needed*/
-    virtual int vMap(double* parameters, int n) = 0;
+    virtual size_t vMap(double* parameters, size_t n) = 0;
 private:
     IGLDrawWork(const IGLDrawWork&) = default;
     IGLDrawWork& operator=(const IGLDrawWork&) = default;
